feat: reports model

+++ db/migrate/20230711044530_create_menu_items.rb
@@ -2,7 +2,7 @@ class CreateMenuItems < ActiveRecord::Migration[7.0]
   def change
     create_table :menu_items do |t|
       t.string :name, null: false
-      t.integer :price, null: false
+      t.decimal :price, null: false
       t.boolean :availability, default: true
 
       t.references :menu_category, null: false, foreign_key: true
